Extract SPI into separate module
